% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptconcept2_sFFLHD_R6_desfuncs.R
\name{actual_des_func_grad_norm2_mean_branin}
\alias{actual_des_func_grad_norm2_mean_branin}
\title{Actual desirability function for the Branin function}
\usage{
actual_des_func_grad_norm2_mean_branin(XX, mod)
}
\arguments{
\item{XX}{Points to evaluate at}

\item{mod}{GP model, not actually used}
}
\description{
Actual desirability function for the Branin function
}
