% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptconcept2_sFFLHD_R6_desfuncs.R
\name{actual_intwerror_func_relmax}
\alias{actual_intwerror_func_relmax}
\title{A func that calculated the actual value of int (weight * |y-hat|) dx}
\usage{
actual_intwerror_func_relmax(mod, alpha, f, fmin, fmax)
}
\arguments{
\item{mod}{IGP model}

\item{alpha}{Parameter to weight desirability}

\item{f}{function that can be evaluted}

\item{fmin}{Min value of function}

\item{fmax}{Max value of function}
}
\description{
A func that calculated the actual value of int (weight * |y-hat|) dx
}
