% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_adaptconceptR6.R
\name{compare.adaptR6}
\alias{compare.adaptR6}
\title{Compare adapt R6}
\description{
Compare adapt R6

Compare adapt R6
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{compare.adaptR6$new()}}
\item \href{#method-set_folder_name}{\code{compare.adaptR6$set_folder_name()}}
\item \href{#method-create_output_folder}{\code{compare.adaptR6$create_output_folder()}}
\item \href{#method-run_all}{\code{compare.adaptR6$run_all()}}
\item \href{#method-run_one}{\code{compare.adaptR6$run_one()}}
\item \href{#method-add_result_of_one}{\code{compare.adaptR6$add_result_of_one()}}
\item \href{#method-postprocess_outdf}{\code{compare.adaptR6$postprocess_outdf()}}
\item \href{#method-plot_MSE_over_batch}{\code{compare.adaptR6$plot_MSE_over_batch()}}
\item \href{#method-plot_AWE_over_batch}{\code{compare.adaptR6$plot_AWE_over_batch()}}
\item \href{#method-plot_AWE_over_group}{\code{compare.adaptR6$plot_AWE_over_group()}}
\item \href{#method-plot_MSE_PVar}{\code{compare.adaptR6$plot_MSE_PVar()}}
\item \href{#method-plot_RMSE_PRMSE}{\code{compare.adaptR6$plot_RMSE_PRMSE()}}
\item \href{#method-plot}{\code{compare.adaptR6$plot()}}
\item \href{#method-plot_run_times}{\code{compare.adaptR6$plot_run_times()}}
\item \href{#method-save_self}{\code{compare.adaptR6$save_self()}}
\item \href{#method-delete_save_folder_if_empty}{\code{compare.adaptR6$delete_save_folder_if_empty()}}
\item \href{#method-recover_parallel_temp_save}{\code{compare.adaptR6$recover_parallel_temp_save()}}
\item \href{#method-parallel_efficiency}{\code{compare.adaptR6$parallel_efficiency()}}
\item \href{#method-clone}{\code{compare.adaptR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$new(
  func,
  D,
  L,
  b = NULL,
  batches = 10,
  reps = 5,
  obj = c("nonadapt", "grad"),
  force_old = c(0),
  force_pvar = c(0),
  n0 = 0,
  stage1batches = 0,
  save_output = F,
  func_string = NULL,
  seed_start = as.numeric(Sys.time()),
  design_seed_start = as.numeric(Sys.time()),
  package = "laGP",
  selection_method = "SMED",
  design = "sFFLHD",
  des_func = NA,
  alpha_des = 1,
  weight_const = 0,
  error_power = 1,
  actual_des_func = NULL,
  pass_list = list(),
  folder_name,
  parallel = FALSE,
  parallel_cores = "detect"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_folder_name"></a>}}
\if{latex}{\out{\hypertarget{method-set_folder_name}{}}}
\subsection{Method \code{set_folder_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$set_folder_name(folder_name, add_timestamp = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_output_folder"></a>}}
\if{latex}{\out{\hypertarget{method-create_output_folder}{}}}
\subsection{Method \code{create_output_folder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$create_output_folder(add_timestamp = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all"></a>}}
\if{latex}{\out{\hypertarget{method-run_all}{}}}
\subsection{Method \code{run_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$run_all(
  redo = FALSE,
  noplot = FALSE,
  save_every = FALSE,
  run_order,
  parallel = self$parallel,
  parallel_temp_save = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_one"></a>}}
\if{latex}{\out{\hypertarget{method-run_one}{}}}
\subsection{Method \code{run_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$run_one(
  irow = NULL,
  save_output = self$save_output,
  noplot = FALSE,
  is_parallel = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_result_of_one"></a>}}
\if{latex}{\out{\hypertarget{method-add_result_of_one}{}}}
\subsection{Method \code{add_result_of_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$add_result_of_one(irow, newdf1, save_output = self$save_output)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-postprocess_outdf"></a>}}
\if{latex}{\out{\hypertarget{method-postprocess_outdf}{}}}
\subsection{Method \code{postprocess_outdf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$postprocess_outdf(save_output = self$save_output)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_MSE_over_batch"></a>}}
\if{latex}{\out{\hypertarget{method-plot_MSE_over_batch}{}}}
\subsection{Method \code{plot_MSE_over_batch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot_MSE_over_batch(
  save_output = self$save_output,
  legend_labels = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_AWE_over_batch"></a>}}
\if{latex}{\out{\hypertarget{method-plot_AWE_over_batch}{}}}
\subsection{Method \code{plot_AWE_over_batch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot_AWE_over_batch(save_output = self$save_output)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_AWE_over_group"></a>}}
\if{latex}{\out{\hypertarget{method-plot_AWE_over_group}{}}}
\subsection{Method \code{plot_AWE_over_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot_AWE_over_group(
  save_output = self$save_output,
  boxpl = TRUE,
  logy = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_MSE_PVar"></a>}}
\if{latex}{\out{\hypertarget{method-plot_MSE_PVar}{}}}
\subsection{Method \code{plot_MSE_PVar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot_MSE_PVar(save_output = self$save_output)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_RMSE_PRMSE"></a>}}
\if{latex}{\out{\hypertarget{method-plot_RMSE_PRMSE}{}}}
\subsection{Method \code{plot_RMSE_PRMSE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot_RMSE_PRMSE(save_output = self$save_output)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot(save_output = self$save_output)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_run_times"></a>}}
\if{latex}{\out{\hypertarget{method-plot_run_times}{}}}
\subsection{Method \code{plot_run_times()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot_run_times()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_self"></a>}}
\if{latex}{\out{\hypertarget{method-save_self}{}}}
\subsection{Method \code{save_self()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$save_self(object_name = "object")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_save_folder_if_empty"></a>}}
\if{latex}{\out{\hypertarget{method-delete_save_folder_if_empty}{}}}
\subsection{Method \code{delete_save_folder_if_empty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$delete_save_folder_if_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recover_parallel_temp_save"></a>}}
\if{latex}{\out{\hypertarget{method-recover_parallel_temp_save}{}}}
\subsection{Method \code{recover_parallel_temp_save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$recover_parallel_temp_save(save_if_any_recovered = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parallel_efficiency"></a>}}
\if{latex}{\out{\hypertarget{method-parallel_efficiency}{}}}
\subsection{Method \code{parallel_efficiency()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$parallel_efficiency()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
