% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptconcept2_sFFLHD_R6_desfuncs.R
\name{des_func_grad_norm2_mean}
\alias{des_func_grad_norm2_mean}
\title{Test desirability functions
A des func for finding large gradient}
\usage{
des_func_grad_norm2_mean(mod, XX, return_se = F)
}
\arguments{
\item{mod}{GauPro model}

\item{XX}{Points to evaluate at}

\item{return_se}{whether the se prediction should be returned along with
the des, all will be returned in data.frame, this will save
time if calculating the werror function since it is faster
to predict both at once instead of separately}
}
\description{
Test desirability functions
A des func for finding large gradient
}
