% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_design.R
\name{sobol_design}
\alias{sobol_design}
\title{Sobol sequence}
\description{
I don't like the randtoolbox because you can't have two
separate concurrent Sobol sequences.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{sobol_design$new()}}
\item \href{#method-get.batch}{\code{sobol_design$get.batch()}}
\item \href{#method-clone}{\code{sobol_design$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sobol_design$new(D, L, seed = numeric(0))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.batch"></a>}}
\if{latex}{\out{\hypertarget{method-get.batch}{}}}
\subsection{Method \code{get.batch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sobol_design$get.batch(L = self$L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{L}}{Number of points to get}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sobol_design$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
