% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptconcept2_sFFLHD_R6_desfuncs.R
\name{test_des_func_grad_norm2_mean}
\alias{test_des_func_grad_norm2_mean}
\title{test_des_func_grad_norm2_mean}
\usage{
test_des_func_grad_norm2_mean(func, actual, d, n = 1000)
}
\arguments{
\item{func}{Function}

\item{actual}{Actual function}

\item{d}{Number of dimensions}

\item{n}{Number of points}
}
\description{
Evaluate a function compared to the actual
using the gradient norm squared of the mean
}
